secure api, edit auctions

+++ pages/api/user/deleteAccount.js
@@ -4,12 +4,12 @@ import dbConnect from "app/lib/dbConnect";
 import { getSession } from "next-auth/react";
 
 export default async (req, res) => {
-  if (req.method === "DELETE") {
-    const session = await getSession({ req });
+  const session = await getSession({ req });
+  if (req.method === "DELETE" && session) {
     await dbConnect();
     const user = await User.findByIdAndDelete(session.user.uid);
     return res.status(200).json({ code: "ok" });
   } else {
-    return res.status(402).json({ err: "WRONG METHOD" });
+    return res.status(405).json({ err: "WRONG METHOD" });
   }
 };
